refactor JdbcProperties.java
